+++ test/chart_engine_test.dart
@@ -1,13 +1,46 @@
-import 'package:chart_engine/chart_engine.dart';
+import 'package:chart_engine/src/chart_engine_colors.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('A group of tests', () {
+  group('ColorPalette', () {
 
     setUp(() {});
 
-    test('First Test', () {
-      //expect(awesome.isAwesome, isTrue);
+    test('Basic palette', () {
+
+      var colorPalette = ColorPalette.from( ['#ff0000','#00ff00','rgba(0,0,255, 0.5)'] ) ;
+
+      expect(colorPalette.basicColors[0], equals( HTMLColor(255,0,0) ));
+      expect(colorPalette.basicColors[1], equals( HTMLColor(0,255,0) ));
+      expect(colorPalette.basicColors[2], equals( HTMLColor(0,0,255, 0.5) ));
+
+      expect(colorPalette.generatePalette(1), equals( [HTMLColor(255,0,0)] ));
+      expect(colorPalette.generatePalette(2), equals( [HTMLColor(255,0,0), HTMLColor(0,255,0)] ));
+      expect(colorPalette.generatePalette(3), equals( [HTMLColor(255,0,0), HTMLColor(0,255,0), HTMLColor(0,0,255, 0.5)] ));
+
     });
+
+    test('Generate palette', () {
+
+      var colorPalette = ColorPalette.from( ['#ff0000','#00ff00'] ) ;
+
+      expect(colorPalette.generateHTMLPalette(3), equals( ['#ff0000', '#00ff00', '#a80000'] ));
+      expect(colorPalette.generateHTMLPalette(4), equals( ['#ff6161', '#ff0000', '#00ff00', '#a80000'] ));
+      expect(colorPalette.generateHTMLPalette(5), equals( ['#ff6161', '#ff0000', '#00ff00', '#a80000', '#00a800'] ));
+      expect(colorPalette.generateHTMLPalette(6), equals( ['#ff6161', '#61ff61', '#ff0000', '#00ff00', '#a80000', '#00a800'] ));
+      expect(colorPalette.generateHTMLPalette(7), equals( ['#ff6161', '#61ff61', '#ff0000', '#00ff00', '#a80000', '#00a800', '#510000'] ));
+      expect(colorPalette.generateHTMLPalette(8), equals( [
+        '#ffc2c2',
+        '#ff6161',
+        '#61ff61',
+        '#ff0000',
+        '#00ff00',
+        '#a80000',
+        '#00a800',
+        '#510000'
+      ] ));
+
+    });
+
   });
 }